+++ ixgbe_82599.h
@@ -13,10 +13,10 @@ struct ixgbe_pipe;
 
 // Returns the new length of the packet, and sets send_list[N] = true iff the packet should be sent on queue N
 // (queues are in the order they were added)
-typedef uint64_t ixgbe_packet_handler(uint8_t* packet, uint64_t packet_length, bool* send_list);
+typedef uint16_t ixgbe_packet_handler(uint8_t* packet, uint16_t packet_length, bool* send_list);
 
 uint64_t ixgbe_get_ring_size(void);
-uint64_t ixgbe_get_packet_size_max(void);
+uint16_t ixgbe_get_packet_size_max(void);
 
 bool ixgbe_device_init(struct tn_pci_device pci_device, struct ixgbe_device** out_device);
 bool ixgbe_device_set_promiscuous(const struct ixgbe_device* device);
